ref(metrics): Isolate minimetrics
