+++ src/minimetrics/__init__.py
@@ -1,0 +1,3 @@
+from core import MetricTagsExternal, MiniMetricsClient
+
+__all__ = ["MiniMetricsClient", "MetricTagsExternal"]
